--- cumulus/parachains/testnets-common/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "testnets-common"
-version = "1.0.0"
-authors.workspace = true
-edition.workspace = true
-description = "Logic and configuration specific to testnet parachain runtimes"
-license = "Apache-2.0"
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-smallvec = "1.11.0"
-
-# Substrate
-frame-support = { path = "../../../substrate/frame/support", default-features = false }
-sp-runtime = { path = "../../../substrate/primitives/runtime", default-features = false }
-
-# Polkadot
-rococo-runtime-constants = { path = "../../../polkadot/runtime/rococo/constants", default-features = false }
-westend-runtime-constants = { path = "../../../polkadot/runtime/westend/constants", default-features = false }
-polkadot-core-primitives = { path = "../../../polkadot/core-primitives", default-features = false }
-
-# Cumulus
-
-[dev-dependencies]
-
-[build-dependencies]
-substrate-wasm-builder = { path = "../../../substrate/utils/wasm-builder" }
-
-[features]
-default = ["std"]
-std = [
-	"frame-support/std",
-	"polkadot-core-primitives/std",
-	"rococo-runtime-constants/std",
-	"sp-runtime/std",
-	"westend-runtime-constants/std",
-]
-
-runtime-benchmarks = [
-	"frame-support/runtime-benchmarks",
-	"sp-runtime/runtime-benchmarks",
-]
